+++ r5dev/game/client/vscript_client.cpp
@@ -12,17 +12,17 @@
 #include "vpc/keyvalues.h"
 #include "engine/cmodel_bsp.h"
 #include "engine/host_state.h"
+#include "engine/client/cl_main.h"
 #include "networksystem/pylon.h"
 #include "networksystem/listmanager.h"
 #include "game/shared/vscript_shared.h"
 #include "vscript/languages/squirrel_re/include/sqvm.h"
 
+#include "vscript_client.h"
+
 namespace VScriptCode
 {
     namespace Client
-    {
-    }
-    namespace Ui
     {
         //-----------------------------------------------------------------------------
         // Purpose: refreshes the server list
@@ -33,6 +33,68 @@ namespace VScriptCode
             size_t iCount = g_pServerListManager->RefreshServerList(serverMessage);
 
             sq_pushinteger(v, static_cast<SQInteger>(iCount));
+
+            return SQ_OK;
+        }
+
+        //-----------------------------------------------------------------------------
+        // Purpose: get current server count from pylon
+        //-----------------------------------------------------------------------------
+        SQRESULT GetServerCount(HSQUIRRELVM v)
+        {
+            size_t iCount = g_pServerListManager->m_vServerList.size();
+            sq_pushinteger(v, static_cast<SQInteger>(iCount));
+
+            return SQ_OK;
+        }
+
+        //-----------------------------------------------------------------------------
+        // Purpose: get response from private server request
+        //-----------------------------------------------------------------------------
+        SQRESULT GetHiddenServerName(HSQUIRRELVM v)
+        {
+            SQChar* privateToken = sq_getstring(v, 1);
+
+            if (!VALID_CHARSTAR(privateToken))
+                return SQ_OK;
+
+            string hiddenServerRequestMessage;
+            NetGameServer_t serverListing;
+
+            bool result = g_pMasterServer->GetServerByToken(serverListing, hiddenServerRequestMessage, privateToken); // Send token connect request.
+            if (!result)
+            {
+                if (hiddenServerRequestMessage.empty())
+                {
+                    sq_pushstring(v, "Request failed", -1);
+                }
+                else
+                {
+                    hiddenServerRequestMessage = Format("Request failed: %s", hiddenServerRequestMessage.c_str());
+                    sq_pushstring(v, hiddenServerRequestMessage.c_str(), -1);
+                }
+
+                return SQ_OK;
+            }
+
+            if (serverListing.m_svHostName.empty())
+            {
+                if (hiddenServerRequestMessage.empty())
+                {
+                    hiddenServerRequestMessage = Format("Server listing empty");
+                }
+                else
+                {
+                    hiddenServerRequestMessage = Format("Server listing empty: %s", hiddenServerRequestMessage.c_str());
+                }
+
+                sq_pushstring(v, hiddenServerRequestMessage.c_str(), -1);
+            }
+            else
+            {
+                hiddenServerRequestMessage = Format("Found server: %s", serverListing.m_svHostName.c_str());
+                sq_pushstring(v, hiddenServerRequestMessage.c_str(), -1);
+            }
 
             return SQ_OK;
         }
@@ -170,17 +232,6 @@ namespace VScriptCode
         }
 
         //-----------------------------------------------------------------------------
-        // Purpose: get current server count from pylon
-        //-----------------------------------------------------------------------------
-        SQRESULT GetServerCount(HSQUIRRELVM v)
-        {
-            size_t iCount = g_pServerListManager->m_vServerList.size();
-            sq_pushinteger(v, static_cast<SQInteger>(iCount));
-
-            return SQ_OK;
-        }
-
-        //-----------------------------------------------------------------------------
         // Purpose: get promo data for serverbrowser panels
         //-----------------------------------------------------------------------------
         SQRESULT GetPromoData(HSQUIRRELVM v)
@@ -239,47 +290,6 @@ namespace VScriptCode
 
             sq_pushstring(v, pszPromoKey, -1);
             return SQ_OK;
-        }
-
-        //-----------------------------------------------------------------------------
-        // Purpose: create server via native serverbrowser entries
-        // TODO: return a boolean on failure instead of raising an error, so we could
-        // determine from scripts whether or not to spin a local server, or connect
-        // to a dedicated server (for disconnecting and loading the lobby, for example)
-        //-----------------------------------------------------------------------------
-        SQRESULT CreateServer(HSQUIRRELVM v)
-        {
-#ifndef CLIENT_DLL
-            SQChar* serverName = sq_getstring(v, 1);
-            SQChar* serverDescription = sq_getstring(v, 2);
-            SQChar* serverMapName = sq_getstring(v, 3);
-            SQChar* serverPlaylist = sq_getstring(v, 4);
-            EServerVisibility_t eServerVisibility = static_cast<EServerVisibility_t>(sq_getinteger(v, 5));
-
-            if (!VALID_CHARSTAR(serverName) ||
-                !VALID_CHARSTAR(serverMapName) ||
-                !VALID_CHARSTAR(serverPlaylist))
-            {
-                return SQ_OK;
-            }
-
-            // Adjust browser settings.
-            std::lock_guard<std::mutex> l(g_pServerListManager->m_Mutex);
-
-            g_pServerListManager->m_Server.m_svHostName = serverName;
-            g_pServerListManager->m_Server.m_svDescription = serverDescription;
-            g_pServerListManager->m_Server.m_svHostMap = serverMapName;
-            g_pServerListManager->m_Server.m_svPlaylist = serverPlaylist;
-            g_pServerListManager->m_ServerVisibility = eServerVisibility;
-
-            // Launch server.
-            g_pServerListManager->LaunchServer();
-
-            return SQ_OK;
-#else
-            v_SQVM_RaiseError(v, "\"%s\" is not supported for client builds.\n", "CreateServer");
-            return SQ_ERROR;
-#endif
         }
 
         //-----------------------------------------------------------------------------
@@ -350,53 +360,11 @@ namespace VScriptCode
         }
 
         //-----------------------------------------------------------------------------
-        // Purpose: get response from private server request
+        // Purpose: checks whether this SDK build is a client dll
         //-----------------------------------------------------------------------------
-        SQRESULT GetHiddenServerName(HSQUIRRELVM v)
+        SQRESULT IsClientDLL(HSQUIRRELVM v)
         {
-            SQChar* privateToken = sq_getstring(v, 1);
-
-            if (!VALID_CHARSTAR(privateToken))
-                return SQ_OK;
-
-            string hiddenServerRequestMessage;
-            NetGameServer_t serverListing;
-
-            bool result = g_pMasterServer->GetServerByToken(serverListing, hiddenServerRequestMessage, privateToken); // Send token connect request.
-            if (!result)
-            {
-                if (hiddenServerRequestMessage.empty())
-                {
-                    sq_pushstring(v, "Request failed", -1);
-                }
-                else
-                {
-                    hiddenServerRequestMessage = Format("Request failed: %s", hiddenServerRequestMessage.c_str());
-                    sq_pushstring(v, hiddenServerRequestMessage.c_str(), -1);
-                }
-
-                return SQ_OK;
-            }
-
-            if (serverListing.m_svHostName.empty())
-            {
-                if (hiddenServerRequestMessage.empty())
-                {
-                    hiddenServerRequestMessage = Format("Server listing empty");
-                }
-                else
-                {
-                    hiddenServerRequestMessage = Format("Server listing empty: %s", hiddenServerRequestMessage.c_str());
-                }
-
-                sq_pushstring(v, hiddenServerRequestMessage.c_str(), -1);
-            }
-            else
-            {
-                hiddenServerRequestMessage = Format("Found server: %s", serverListing.m_svHostName.c_str());
-                sq_pushstring(v, hiddenServerRequestMessage.c_str(), -1);
-            }
-
+            sq_pushbool(v, ::IsClientDLL());
             return SQ_OK;
         }
     }
@@ -408,14 +376,8 @@ namespace VScriptCode
 //---------------------------------------------------------------------------------
 void Script_RegisterClientFunctions(CSquirrelVM* s)
 {
-    s->RegisterFunction("SDKNativeTest", "Script_SDKNativeTest", "Native CLIENT test function", "void", "", &VScriptCode::Shared::SDKNativeTest);
-    s->RegisterFunction("GetSDKVersion", "Script_GetSDKVersion", "Gets the SDK version as a string", "string", "", &VScriptCode::Shared::GetSDKVersion);
-
-    s->RegisterFunction("GetAvailableMaps", "Script_GetAvailableMaps", "Gets an array of all available maps", "array< string >", "", &VScriptCode::Shared::GetAvailableMaps);
-    s->RegisterFunction("GetAvailablePlaylists", "Script_GetAvailablePlaylists", "Gets an array of all available playlists", "array< string >", "", &VScriptCode::Shared::GetAvailablePlaylists);
-
-    s->RegisterFunction("ShutdownHostGame", "Script_ShutdownHostGame", "Shuts the local host game down", "void", "", &VScriptCode::Shared::ShutdownHostGame);
-    s->RegisterFunction("IsClientDLL", "Script_IsClientDLL", "Returns whether this build is client only", "bool", "", &VScriptCode::Shared::IsClientDLL);
+    Script_RegisterCommonAbstractions(s);
+    Script_RegisterCoreClientFunctions(s);
 }
 
 //---------------------------------------------------------------------------------
@@ -424,47 +386,37 @@ void Script_RegisterClientFunctions(CSquirrelVM* s)
 //---------------------------------------------------------------------------------
 void Script_RegisterUIFunctions(CSquirrelVM* s)
 {
-    s->RegisterFunction("SDKNativeTest", "Script_SDKNativeTest", "Native UI test function", "void", "", &VScriptCode::Shared::SDKNativeTest);
-    s->RegisterFunction("GetSDKVersion", "Script_GetSDKVersion", "Gets the SDK version as a string", "string", "", &VScriptCode::Shared::GetSDKVersion);
+    Script_RegisterCommonAbstractions(s);
+    Script_RegisterCoreClientFunctions(s);
 
-    s->RegisterFunction("RefreshServerList", "Script_RefreshServerList", "Refreshes the public server list and returns the count", "int", "", &VScriptCode::Ui::RefreshServerCount);
+    s->RegisterFunction("RefreshServerList", "Script_RefreshServerList", "Refreshes the public server list and returns the count", "int", "", &VScriptCode::Client::RefreshServerCount);
+    s->RegisterFunction("GetServerCount", "Script_GetServerCount", "Gets the number of public servers", "int", "", &VScriptCode::Client::GetServerCount);
 
     // Functions for retrieving server browser data
-    s->RegisterFunction("GetServerName", "Script_GetServerName", "Gets the name of the server at the specified index of the server list", "string", "int", &VScriptCode::Ui::GetServerName);
-    s->RegisterFunction("GetServerDescription", "Script_GetServerDescription", "Gets the description of the server at the specified index of the server list", "string", "int", &VScriptCode::Ui::GetServerDescription);
-    s->RegisterFunction("GetServerMap", "Script_GetServerMap", "Gets the map of the server at the specified index of the server list", "string", "int", &VScriptCode::Ui::GetServerMap);
-    s->RegisterFunction("GetServerPlaylist", "Script_GetServerPlaylist", "Gets the playlist of the server at the specified index of the server list", "string", "int", &VScriptCode::Ui::GetServerPlaylist);
-    s->RegisterFunction("GetServerCurrentPlayers", "Script_GetServerCurrentPlayers", "Gets the current player count of the server at the specified index of the server list", "int", "int", &VScriptCode::Ui::GetServerCurrentPlayers);
-    s->RegisterFunction("GetServerMaxPlayers", "Script_GetServerMaxPlayers", "Gets the max player count of the server at the specified index of the server list", "int", "int", &VScriptCode::Ui::GetServerMaxPlayers);
-    s->RegisterFunction("GetServerCount", "Script_GetServerCount", "Gets the number of public servers", "int", "", &VScriptCode::Ui::GetServerCount);
+    s->RegisterFunction("GetHiddenServerName", "Script_GetHiddenServerName", "Gets hidden server name by token", "string", "string", &VScriptCode::Client::GetHiddenServerName);
+    s->RegisterFunction("GetServerName", "Script_GetServerName", "Gets the name of the server at the specified index of the server list", "string", "int", &VScriptCode::Client::GetServerName);
+    s->RegisterFunction("GetServerDescription", "Script_GetServerDescription", "Gets the description of the server at the specified index of the server list", "string", "int", &VScriptCode::Client::GetServerDescription);
+
+    s->RegisterFunction("GetServerMap", "Script_GetServerMap", "Gets the map of the server at the specified index of the server list", "string", "int", &VScriptCode::Client::GetServerMap);
+    s->RegisterFunction("GetServerPlaylist", "Script_GetServerPlaylist", "Gets the playlist of the server at the specified index of the server list", "string", "int", &VScriptCode::Client::GetServerPlaylist);
+    s->RegisterFunction("GetServerCurrentPlayers", "Script_GetServerCurrentPlayers", "Gets the current player count of the server at the specified index of the server list", "int", "int", &VScriptCode::Client::GetServerCurrentPlayers);
+
+    s->RegisterFunction("GetServerMaxPlayers", "Script_GetServerMaxPlayers", "Gets the max player count of the server at the specified index of the server list", "int", "int", &VScriptCode::Client::GetServerMaxPlayers);
 
     // Misc main menu functions
-    s->RegisterFunction("GetPromoData", "Script_GetPromoData", "Gets promo data for specified slot type", "string", "int", &VScriptCode::Ui::GetPromoData);
-
-    // Functions for creating servers
-    s->RegisterFunction("CreateServer", "Script_CreateServer", "Starts server with the specified settings", "void", "string, string, string, string, int", &VScriptCode::Ui::CreateServer);
-    s->RegisterFunction("IsServerActive", "Script_IsServerActive", "Returns whether the server is active", "bool", "", &VScriptCode::Shared::IsServerActive);
+    s->RegisterFunction("GetPromoData", "Script_GetPromoData", "Gets promo data for specified slot type", "string", "int", &VScriptCode::Client::GetPromoData);
 
     // Functions for connecting to servers
-    s->RegisterFunction("ConnectToServer", "Script_ConnectToServer", "Joins server by ip address and encryption key", "void", "string, string", &VScriptCode::Ui::ConnectToServer);
-    s->RegisterFunction("ConnectToListedServer", "Script_ConnectToListedServer", "Joins listed server by index", "void", "int", &VScriptCode::Ui::ConnectToListedServer);
-    s->RegisterFunction("ConnectToHiddenServer", "Script_ConnectToHiddenServer", "Joins hidden server by token", "void", "string", &VScriptCode::Ui::ConnectToHiddenServer);
+    s->RegisterFunction("ConnectToServer", "Script_ConnectToServer", "Joins server by ip address and encryption key", "void", "string, string", &VScriptCode::Client::ConnectToServer);
+    s->RegisterFunction("ConnectToListedServer", "Script_ConnectToListedServer", "Joins listed server by index", "void", "int", &VScriptCode::Client::ConnectToListedServer);
+    s->RegisterFunction("ConnectToHiddenServer", "Script_ConnectToHiddenServer", "Joins hidden server by token", "void", "string", &VScriptCode::Client::ConnectToHiddenServer);
+}
 
-    s->RegisterFunction("GetHiddenServerName", "Script_GetHiddenServerName", "Gets hidden server name by token", "string", "string", &VScriptCode::Ui::GetHiddenServerName);
-    s->RegisterFunction("GetAvailableMaps", "Script_GetAvailableMaps", "Gets an array of all available maps", "array< string >", "", &VScriptCode::Shared::GetAvailableMaps);
-    s->RegisterFunction("GetAvailablePlaylists", "Script_GetAvailablePlaylists", "Gets an array of all available playlists", "array< string >", "", &VScriptCode::Shared::GetAvailablePlaylists);
-
-#ifndef CLIENT_DLL // UI 'admin' functions controlling server code
-    s->RegisterFunction("KickPlayerByName", "Script_KickPlayerByName", "Kicks a player from the server by name", "void", "string", &VScriptCode::SHARED::KickPlayerByName);
-    s->RegisterFunction("KickPlayerById", "Script_KickPlayerById", "Kicks a player from the server by handle or nucleus id", "void", "string", &VScriptCode::SHARED::KickPlayerById);
-
-    s->RegisterFunction("BanPlayerByName", "Script_BanPlayerByName", "Bans a player from the server by name", "void", "string", &VScriptCode::SHARED::BanPlayerByName);
-    s->RegisterFunction("BanPlayerById", "Script_BanPlayerById", "Bans a player from the server by handle or nucleus id", "void", "string", &VScriptCode::SHARED::BanPlayerById);
-
-    s->RegisterFunction("UnbanPlayer", "Script_UnbanPlayer", "Unbans a player from the server by nucleus id or ip address", "void", "string", &VScriptCode::SHARED::UnbanPlayer);
-
-    s->RegisterFunction("ShutdownHostGame", "Script_ShutdownHostGame", "Shuts the local host game down", "void", "", &VScriptCode::SHARED::ShutdownHostGame);
-#endif // !CLIENT_DLL
-
-    s->RegisterFunction("IsClientDLL", "Script_IsClientDLL", "Returns whether this build is client only", "bool", "", &VScriptCode::Shared::IsClientDLL);
+//---------------------------------------------------------------------------------
+// Purpose: core client script functions
+// Input  : *s - 
+//---------------------------------------------------------------------------------
+void Script_RegisterCoreClientFunctions(CSquirrelVM* s)
+{
+    s->RegisterFunction("IsClientDLL", "Script_IsClientDLL", "Returns whether this build is client only", "bool", "", &VScriptCode::Client::IsClientDLL);
 }
