+++ r5dev/vscript/languages/squirrel_re/vsquirrel.cpp
@@ -13,6 +13,10 @@ void(*ServerScriptRegister_Callback)(CSquirrelVM* s) = nullptr;
 void(*ClientScriptRegister_Callback)(CSquirrelVM* s) = nullptr;
 void(*UiScriptRegister_Callback)(CSquirrelVM* s) = nullptr;
 
+// Admin panel functions, NULL on client only builds.
+void(*CoreServerScriptRegister_Callback)(CSquirrelVM* s) = nullptr;
+void(*AdminPanelScriptRegister_Callback)(CSquirrelVM* s) = nullptr;
+
 //---------------------------------------------------------------------------------
 // Purpose: Initialises a Squirrel VM instance
 // Output : True on success, false on failure
@@ -29,28 +33,32 @@ SQBool CSquirrelVM::Init(CSquirrelVM* s, SQCONTEXT context, SQFloat curTime)
 
 	switch (context)
 	{
-#ifndef CLIENT_DLL
 	case SQCONTEXT::SERVER:
 		g_pServerScript = s;
+
 		if (ServerScriptRegister_Callback)
 			ServerScriptRegister_Callback(s);
 
 		break;
-#endif
-#ifndef DEDICATED
 	case SQCONTEXT::CLIENT:
 		g_pClientScript = s;
+
 		if (ClientScriptRegister_Callback)
 			ClientScriptRegister_Callback(s);
 
 		break;
 	case SQCONTEXT::UI:
 		g_pUIScript = s;
+
 		if (UiScriptRegister_Callback)
 			UiScriptRegister_Callback(s);
 
+		if (CoreServerScriptRegister_Callback)
+			CoreServerScriptRegister_Callback(s);
+		if (AdminPanelScriptRegister_Callback)
+			AdminPanelScriptRegister_Callback(s);
+
 		break;
-#endif
 	}
 
 	return true;
@@ -110,21 +118,17 @@ void CSquirrelVM::SetAsCompiler(RSON::Node_t* rson)
 	const SQCONTEXT context = GetContext();
 	switch (context)
 	{
-#ifndef CLIENT_DLL
 	case SQCONTEXT::SERVER:
 	{
 		v_Script_SetServerPrecompiler(context, rson);
 		break;
 	}
-#endif
-#ifndef DEDICATED
 	case SQCONTEXT::CLIENT:
 	case SQCONTEXT::UI:
 	{
 		v_Script_SetClientPrecompiler(context, rson);
 		break;
 	}
-#endif
 	}
 }
 
@@ -187,21 +191,17 @@ void CSquirrelVM::CompileModScripts()
 
 			switch (GetVM()->GetContext())
 			{
-#ifndef CLIENT_DLL
 			case SQCONTEXT::SERVER:
 			{
 				v_CSquirrelVM_PrecompileServerScripts(this, GetContext(), (char**)scriptPathArray, scriptCount);
 				break;
 			}
-#endif
-#ifndef DEDICATED
 			case SQCONTEXT::CLIENT:
 			case SQCONTEXT::UI:
 			{
 				v_CSquirrelVM_PrecompileClientScripts(this, GetContext(), (char**)scriptPathArray, scriptCount);
 				break;
 			}
-#endif
 			}
 
 			// clean up our allocated script paths
